refactor: 로또 당첨 기준 상수 분리
