+++ src/constants/Messages.js
@@ -19,11 +19,3 @@ export const ERROR_MESSAGE =  {
     DUPLICATE_NUMBERS: "중복된 로또 번호가 존재합니다.",
     DUPLICATE_WINNING_LOTTO_NUMBERS: "당첨 번호와 중복된 보너스 번호입니다.",
 };
-
-export const WINNING_CRITERIA = {
-    5: { price: 5000, message: "3개 일치 (5,000원)" },
-    4: { price: 50000, message: "4개 일치 (50,000원)" },
-    3: { price: 1500000, message: "5개 일치 (1,500,000원)" },
-    2: { price: 30000000, message: "5개 일치, 보너스 볼 일치 (30,000,000원)" },
-    1: { price: 2000000000, message: "6개 일치 (2,000,000,000원)" },
-};
